+++ pasqal_cloud/errors.py
@@ -1,4 +1,31 @@
-class BatchException(BaseException):
+import json
+from typing import Optional
+
+from requests import HTTPError, Response
+
+
+class ExceptionWithResponseContext(BaseException):
+    def __init__(self, msg: str, e: Optional[HTTPError] = None) -> None:
+        if not e:
+            return super().__init__(msg)
+        resp: Response = e.response
+        if not resp.content:
+            return super().__init__(msg)
+        data = resp.json()
+
+        code = data.get("code", 0)
+
+        if data.get("data", ""):
+            description = data["data"].get("description", data.get("message"))
+        else:
+            description = data.get("message", "")
+
+        super().__init__(
+            f"{msg}: {code}: {description}\nDetails: {json.dumps(data, indent=2)}"
+        )
+
+
+class BatchException(ExceptionWithResponseContext):
     """
     Base Exception class for batches
     """
@@ -9,8 +36,8 @@ class BatchCreationError(BatchException):
     Exception class when batch creation failed
     """
 
-    def __init__(self) -> None:
-        super().__init__("Batch creation failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Batch creation failed", e)
 
 
 class BatchFetchingError(BatchException):
@@ -18,8 +45,8 @@ class BatchFetchingError(BatchException):
     Exception class raised when batch fetching failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Batch fetching failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Batch fetching failed", e)
 
 
 class BatchCancellingError(BatchException):
@@ -27,8 +54,8 @@ class BatchCancellingError(BatchException):
     Exception class raised when batch cancelling failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Batch cancelling failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Batch cancelling failed", e)
 
 
 class BatchSetCompleteError(BatchException):
@@ -36,11 +63,11 @@ class BatchSetCompleteError(BatchException):
     Exception class raised when setting batch to complete failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Batch setting to complete failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Batch setting to complete failed", e)
 
 
-class JobException(BaseException):
+class JobException(ExceptionWithResponseContext):
     """
     Base Exception class for jobs.
     """
@@ -51,14 +78,17 @@ class JobCreationError(JobException):
     Exception class raised when job creation failed.
     """
 
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Job creation failed", e)
+
 
 class JobFetchingError(JobException):
     """
     Exception class raised when job fetching failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Job fetching failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Job fetching failed", e)
 
 
 class JobCancellingError(JobException):
@@ -66,11 +96,11 @@ class JobCancellingError(JobException):
     Exception class raised when job cancelling failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Job cancelling failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Job cancelling failed", e)
 
 
-class WorkloadException(BaseException):
+class WorkloadException(ExceptionWithResponseContext):
     """
     Base exception class for workloads.
     """
@@ -81,8 +111,8 @@ class WorkloadFetchingError(WorkloadException):
     Exception class raised when workload fetching failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Workload fetching failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Workload fetching failed", e)
 
 
 class WorkloadCreationError(WorkloadException):
@@ -90,8 +120,8 @@ class WorkloadCreationError(WorkloadException):
     Exception class raised when workload creation failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Job creation failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Job creation failed", e)
 
 
 class WorkloadCancellingError(WorkloadException):
@@ -99,8 +129,8 @@ class WorkloadCancellingError(WorkloadException):
     Exception class raised when cancelling workload failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Workload cancelling failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Workload cancelling failed", e)
 
 
 class WorkloadResultsDownloadError(WorkloadException):
@@ -108,8 +138,17 @@ class WorkloadResultsDownloadError(WorkloadException):
     Exception class raised when failed to download results.
     """
 
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Workload results download failed.", e)
+
+
+class WorkloadResultsDecodeError(WorkloadException):
+    """
+    Exception class raised when download results succeeded but decoding failed.
+    """
+
     def __init__(self) -> None:
-        super().__init__("Workload results download failed.")
+        super().__init__("Workload results decoding failed.")
 
 
 class InvalidWorkloadResultsFormatError(WorkloadException):
@@ -133,5 +172,5 @@ class DeviceSpecsFetchingError(DeviceSpecsException):
     Exception class raised when fetching of device specs failed.
     """
 
-    def __init__(self) -> None:
-        super().__init__("Device specs fetching failed.")
+    def __init__(self, e: HTTPError) -> None:
+        super().__init__("Device specs fetching failed", e)