+++ CHANGELOG.md
@@ -2,11 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.2] - 2023-10-09
+
+### Added
+
+- Added new base error class to inject response context data in exception message.
+- Added tests for new base error class.
+
+### Changed
+
+- Updated existing error to inherit from new response context exception class.
+
 ## [0.4.1] - 2023-10-09
 
 ### Added
 
-Added `result_link` field to `Workload` object
+- Added `result_link` field to `Workload` object
 
 ### Changed